Update seeder with legacy ID logic
We intended to order the return requirements in the view by return reference (legacy ID behind the scenes).

The reference forms the first part of the title followed by the site description. So, unlike the `/check` page, we need to order by it and not site reference (the reference is only generated when the return version is approved on the `/check` page).

To allow us to test we've got the ordering right, we need to control what references we populate our test records with. But we also have to do it in a way that re-running the tests wouldn't cause an error because we've tried to seed a test record with a duplicate reference.