Create new service to handle 'cancelling'
The key thing we've realised is we need to separate the 'cancelling' a bill run from the deleting of all the data.

So, this is the first step; a new service that extracts out the responsibility for checking if a bill run can be deleted, and if so updating its status.