Create new service to handle 'deleting'
The next step is to extract out all the deleting functionality from the existing `SubmitCancelBillRunService`. Not only does it become easier to manage and test these services, it means we can revert back to more conventional means of promise and error handling in the SubmitCancelBillRunService, even whilst awaiting one step and starting the other in the background.