The refactored lean & mean SubmitCancelBillRun !
CancelBillRunService now just has to deal with cancelling the bill run. DeleteBillRunService is only responsible for deleting the bill run.

This means SubmitCancelBillRun can focus on orchestrating the two, and our tests can focus on the key behaviour we are wanting to elicit: that we await the cancel and let the delete happen in the background. :-)