Migrate licence points view page (#1451)
* Migrate licence points view page

https://eaflood.atlassian.net/browse/WATER-4727

Having migrated the view licence page to our new code base, we are starting to look at the various pages linked from it.

Migrating the legacy pages not only means we can better support the service, but it is also a necessary step to updating to the latest
GDS design system components and then carrying out an accessibility audit. Next on our hit list is the Licence points view page, which is
linked from the summary tab.

This PR is focused on rebuilding the licence points view page from the legacy code base.