Create new service to handle 'sending'
The key thing we've realised is we need to separate the 'sending' a bill run from the updating of all the data.

So, this is the first step; a new service that extracts out the responsibility for checking if a bill run can be sent, and if so updating its status.