Create new service to handle 'updating'
The next step is to extract out all the updating functionality from the existing `SubmitSendBillRunService`. Not only does it become easier to manage and test these services, it means we can revert back to more conventional means of promise and error handling in the SubmitSendBillRunService, even whilst awaiting one step and starting the other in the background.