The refactored lean & mean SubmitSendBillRun !
SendBillRunService now just has to deal with sending the bill run. UpdateInvoiceNumbersService is only responsible for fetching the invoice numbers from the CHA and applying them to the bills.

This means SubmitSendBillRun can focus on orchestrating the two, and our tests can focus on the key behaviour we are wanting to elicit: that we await the send and let the update happen in the background. :-)