Update the controller
Now we sorted out the underlying modules, we no longer have to explicitly handle errors in the controller. If SubmitSendBillRunService throws an error, it will be from SendBillRunService, and because it is awaited it will happen in the same 'loop' (no idea what the correct term is!) Our standard error handling can therefore take care of it.

Anything the UpdateInvoiceNumbersService throws is handled in that service so never bubbles up as an uncaught exception causing the app to crash, which was the reason for the non-conventional error handling prior to this refactor.