Stop returning command-only async calls (#1499)
There are two scenarios where folks will see @Cruikshanks sweep in and ask for a change.

```javascript
// Scenario 1
async function go (someId) {
  const result = await _fetch(someId)

  return result
}

// Scenario 2
async function go (someId) {
  await _updateStatus(someId)
}
```

In both cases, he'll ask folks to return the call immediately.

```javascript
// Scenario 1
async function go (someId) {
  return _fetch(someId)
}

// Scenario 2
async function go (someId) {
  return _updateStatus(someId)
}
```

In a recent change where we added a job, we made a command-only async call as a batch to `Promise.all()`. The original change had `await ProcessLicence.go(licenceId)`, so in swooped @Cruikshanks !

This was challenged, so we carried out a mini spike, did some Googling, and came to the conclusion.

- We have it bang on for scenario 1; we should return immediately
- @Cruikshanks has it _so wrong_ for scenario 2 🤦

The spike concluded that, technically, there is no difference. Using `await` rather than `return` doesn't block anything, and the resolution is the same.

Googling backed up the challenge made: declaring `return _doThing()` makes it look like we intend to return something when we don't. When doing this, we have to add the following to the JSDocs, which should have clued us in earlier.

> `@returns {Promise<object>} the promise returned is not intended to resolve to any particular value`

We found a blog post that put into 'bigger' words (!!) why this is bad.

> `await` should be used for commands to ensure void is returned. The reason why void is returned for command is to enforce [command-query separation](https://en.wikipedia.org/wiki/Command%E2%80%93query_separation). Do not mix updates with reads as that violates [single-responsibility principle](https://en.wikipedia.org/wiki/Single-responsibility_principle).
>
> [When to use return vs await vs return await vs nothing](https://blog.battlefy.com/when-to-use-return-vs-await-vs-return-await-vs-nothing)

So, this change is @Cruikshanks putting all the instances where he got it wrong right!