Fix null values in view monitoring station page
https://eaflood.atlassian.net/browse/WATER-4736

We had some issues in initial version of the new view monitoring station page we built. We [fixed them in](https://github.com/DEFRA/water-abstraction-system/pull/1470).

As part of that change though, we dropped some logic in the presenter that converted `null` to an empty string.

We thought it was unnecessary, but we were wrong.

In Nunjucks, if you render a value using `{{ gridReference }}` Nunjucks will automatically convert `undefined` and `null` to an empty string. This means you don't have to do anything in the presenter.

But, if you render the value as part of some string concatenation this no longer happens and `null` gets converted into `'null'`. A common reason for doing this is adding a data-test attribute to GDS components that don't allow us to add custom attributes.

Note the `'<span data-test="meta-data-grid-reference">' + gridReference + '</span>'` below.

```
  {{ govukSummaryList({
    classes: 'govuk-summary-list--no-border',
    attributes: {
      'data-test': 'meta-data'
    },
    rows: [
      {
        key: { text: "Grid reference", classes: "meta-data__label" },
        value: { html: '<span data-test="meta-data-grid-reference">' + gridReference + '</span>', classes: "meta-data__value" }
      }
    ]
  }) }}
```

This change puts back the logic. It also corrects a misnamed attribute!