Update bill run statuses in DetermineExistingBillRunService
https://eaflood.atlassian.net/browse/WATER-4746

This PR is part of our work to clean up the supplementary billing flags. Specifically this PR focuses on correcting the bill run statuses that are used when we check if we should flag for two-part tariff supplementary billing.
Currently bill run status are not apart of the check, however there are a few statuses such as 'errored' or 'cancelled' that we don't want to consider.