Add determine licence flags service (#1506)
https://eaflood.atlassian.net/browse/WATER-4746

This PR is contributing to the cleaning up of the supplementary billing flags process. This PR specifically focuses on the licence flags service which is legacy code we have migrated to use the new communal supplementary billing flags service. It will flag a licence based on the provided scheme. The flagging mechanisms that will use this service are:

A new licence agreement
An ended licence agreement
A deleted licence agreement
Removing a licence from a pre-sroc two-part tariff bill run.
As this is handling legacy flagging this service only cares about the pre-sroc and sroc flags. Two-part tariff years flagging is not included for any of these so we default it to false.