Add a new page for notifications returns period
https://eaflood.atlassian.net/browse/WATER-4716

As part of ongoing work to migrate the legacy code we are rebuilding the notifications journeys.

This change adds a new journey for notifications. This accessed from the legacy /manage page and is behind a feature toggle.

The initial notification routes and controller scaffolding has been put in place to build the journey.

The returns period page has been scaffolded to facilitate further development of the page and to get an idea of language used, established patterns and code structure.

This change will give us a good foundation for the notifications journey.

The service logic and submit route will be added in a later commit.