Update determine charge versions flags service
https://eaflood.atlassian.net/browse/WATER-4746

This PR contributes to cleaning up the supplementary billing flagging process, specifically focusing on updating the determine charge versions flags service.
Currently the service is equip to handle new charge versions being approved and flagging the licence based on the new charge version. However this doesn't take into account the fact the licence has been in workflow and might need additional flags because it wasn't able to be picked up for billing.
For example:
A new two-part tariff charge version gets approved for 2022-2023 and the licence needs to flag 2022 and 2023 in the licence supplementary years table. This is done successfully with the current service but if the licence then missed its annual bill run because it was in workflow waiting for that charge version to be approved, it would then also need an sroc supplementary flag as well.

This PR fixes that issue.