Al's take on Return log for ended licence
https://eaflood.atlassian.net/browse/WATER-4546

This is going to be an integral and complex part of the service, but until now I've only had a tertiary level of understanding of how it is being pulled together.

In [Return log for ended licence](https://github.com/DEFRA/water-abstraction-system/pull/1425) we're making some fundamental changes and my feedback is being sort, so it is time to get my hands dirty.

There is likely to be a lot of feedback to provide, and when this is the case I find it easier to do so in a separate PR using my commits to highlight and explain my changes. Also allows me to check what I'm suggesting actually works!