Simplify controller and follow conventions
This controller is purely for testing so we don't need to handle if someone gives us an invalid payload. That is on them!

However, if we were it would not be a `404` we should return. That would make more sense if the licence ref provided was unknown. I would have gone with a [400 - bad request](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/400). Finally, if we were doing this checking, it should be in the service and not the controller, to make the controller as 'skinny' as possible, and making it much easier to test.

We can also `await` the call. We're only doing the one licence so _hopefully_ we can complete the work in less than a second! And again, this is for testing purposes only, so its both handy to know the work is actually done, plus we get a sense of how long it is taking (and whether we need to look at performance).

On the convention point, we've _never_ gotten the payload by way of the Hapi response object. I'm not sure why we needed to here.