Service JSDoc description change
> This might not be the 'right' description for this service but I'll know more later

Original had a typo, was longer than 12o characters (so broke our line length convention), ended with a period (sure they'll be an exception but we don't end the opening description with a period anywhere else), and didn't match the module one liner.

These are all out of sync with how JSDoc's are done in other files.