Notifications returns period options
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change adds the notifications returns periods option for between january.

The other options will be added in following commits.

The submit logic will be developed in a separate commit so the 'values' used to capture the selected options may change.