Add determine bill flags service (#1523)
https://eaflood.atlassian.net/browse/WATER-4746

This PR contributes to cleaning up the supplementary billing flagging process, specifically focusing on handling bills removed from a bill run.

When a bill run is created, users can remove bills from it. Depending on the type of bill run, the following actions are required:

Annual Bill Run: The licences in the bill must be flagged for inclusion in the next supplementary bill run.
Supplementary Bill Run: Since the licences in the bill are already flagged (which is why they were included in this bill run), the existing flag must be maintained.

Previously, bill flagging logic was handled entirely in legacy code. While the endpoint in the service repository still processes other aspects of licence removal, the flagging logic has been moved to our new ProcessBillingFlagsService. This service ensures flags are correctly determined before the legacy endpoint is called.