Remove test-only single licence from job code
The addition of the /check endpoint for a single licence and the `app/services/return-logs/process-licence-return-logs.service.js` service means we know longer need a means to check a single licence. The fact `ProcessLicenceReturnLogs` defaults to the current date _should_ result in the same outcome as if the job looked at it.