Refactor returns period date handling
https://eaflood.atlassian.net/browse/WATER-4716

Previous has been done to implement the required output for the work in the ticket.

But we can do more in the code to make this closer to the returns cycles.

This change will refactor the returns period presenter to better align with the workings of return cycles, rather than just fixing dates boundaries with no specific reason / meaning.