Create job specific fetch rtn reqs services
The history of the fetch service is it was written to fetch all return requirements for a cycle. Then we altered it to take an optional licence ref to allow us to test the job more easily.

Now we have a `/check` endpoint that allows us to see what the engine will do for a specific licence. It reused the fetch as it could already take an optional licence ref and amend its result based on that.

I've now removed support for passing a licence ref in the job endpoint as we have the `/check` endpoint. I also believe the fetch in its current form was made unnecessarily complex because of this dual need. If stop sharing the fetch service and create ones specific to their need, I think we'll have a clearer idea of what, if anything, is 'shareable'. I also think it allows us to simplify what we are doing.

So, to start to explore this I have copied the fetch and put the copy back into the jobs folder. I have then removed all functionality related to supporting the optional licence ref.

I've also completely re-written the tests to align with the behaviour and conditions in the service. It has exposed we're fetching a lot of data we're not using. So, next we'll look at what can be refactored.