Move where exists query into main body
We have put the query that will be used in a whereExists / whereNotExists into a separate function in other places. But typically that is because they will be reused within the Fetch module.

In this module that is not the case so we move it back into the main body where it can be seen in context.

> not sure about this move but going with it for now.