Update and repair FetchCurrentReturnCycleService
Following on from our previous change, the service now grabs the current dates and does what it needs to with it. Doing this I've reduced some of the 'date -> to string -> to date -> to string' chain that was going on.

Previously, because the date was being determined outside of the service and passed in, I'm not sure what the Sinon clock was doing in the tests. In some cases it even differed from the date used in the test.

But now we actively depend on it so I ensure it is set to match the previous dates that were being passed in. Other than that we correct the it statements (we are not just passing back a UUID), highlight the key params and values being used, and ensure they are consistent across to two main scenarios.

Back in the service, we update the JSDocs, and get them matching (module -> function). We correct the return type. We also remove a redundant `await`.