Rename as CreateCurrentReturnCycle
Renaming to match conventions and make explicit what it is doing.

When a service adds something to the DB, our convention is to name it `Create`. We tend to use `Generate` for when generating the data to be inserted is complex enough it requires its own service to manage.

> I'm sure there are plenty of instances we are not following this convention, but we should be!

We add `Current` to the service name to make it explicit it creates the requested cycle based on the 'current' date. It also helps to see its partnership with `FetchCurrentReturnCycleService` more clearly.