Update and repair CreateCurrentReturnCycleService
I'm guessing the actual insert query was copied from the seed, or done to avoid errors in the unit tests, hence its an upsert. In reality, we should never be attempting to insert a return cycle which already exists, so better for the service to blow up so we can see when and where this has happened.

We do the usual tidy up of the JSDoc comments. We also directly return the result of the patch instead of saving it to a local variable before returning.

We have a function specific to PostgresSQL timestamps so we use it here for consistency. Any other changes are purely personal preference.

In the tests, they are clearly reliant on the service using an upsert. Refactoring it back to simple insert means I'm comfortable stubbing out the actually database hit. We can still ensure the data it is attempting to insert is as expected. And again, we can show the partnership between this and the FetchCurrentReturnCycleService by mirroring the tests it is using. Sinon.clock allows us to control the date and be explicit about the scenario we are running in our tests.

Finally, as we have replaced the upsert with a standard insert, we add tests to demonstrate the service will error if we attempt to create an existing return cycle.