Use new determineLatestDate() helper in service
Update GenerateReturnLogService to use the new helper. We also ensure licence start date is added into the mix. When the return version is created in WRLS, we'll ensure that its start date cannot be earlier than the licence start date so they'll be the same).

But we're dealing with records created in NALD, and even after we start managing return versions, there is nothing stopping users amending the licence start date in NALD to a new, later date _after_ the return version has been created.

So, we should ensure we are not creating return logs that start before the licence does.