Rewrite the GenerateReturnLogService tests
By re-writing the tests we can resolve a number of issues with the existing tests.

They currently have to alter depending on the date they are run. We can cater for that by using Sinon timers to control what the 'current' date is. There is actually very little logic in the service that is cycle dependent and we can cover those cases in other ways.

They are heavily dependent on using real records committed to the DB. However, it relies purely on the data being passed in. We can therefore generate our own objects and control the services inputs that way.

Previously, the service expected to be passed multiple return requirements so would return multiple generated return logs. Now it's just focused on converting one requirement to a return log (much better!). However, the tests all exclaim '[it] should return one return log payload'. This isn't telling us anything about the behaviour. And always assert against all properties of the object returned. By specifying the behaviours we intend to illicit, we can tailor the scenarios (data inputs) specifically to them and make what has changed much clearer.