Migrate abstraction conditions view page
As the legacy licence view page has now been rebuilt, the next process is to update some of the links pointing to legacy pages.
One of these legacy pages is the abstraction conditions view. If we are able to move this over to our system, it will help towards removing the use of old legacy code.

This PR is focused on rebuilding the abstraction conditions view page from the legacy code base.