A note about possible performance issues
I've dropped a `// NOTE:` about the point in the code I'm referring to.

This is intended to be used in the background over night so its right not to be overly concerned about performance.

But when this runs, at the point in the code I've highlighted we would have the following

- thousands of return requirements plus their associated data held in memory
- thousands of return logs generated from those return requirements held in memory

I ran the query today on the production replica and got 5,564 results back from `FetchReturnRequirementsService`. It's also worth knowing that when iterating through the results, and creating a new array from them, we're essentially asking Node to resize the return logs array 5K times!

So, in the next change I'm going to refactor `ProcessReturnLogs` to reduce the need for storing return logs before inserting them in the DB. This will also set up the loop nicely for being batched (but that is for a future change!)