Refactor ProcessReturnLogsService and tests
As highlighted in the previous comment, ideally we'd avoid accruing the return logs to persist. This change updates the service to generate the return log record then immediately persist it.

This means we are not storing the data in an array, or calling `push()` 5K times. It does mean we no longer have a use for `CreateReturnLogsService` in this service, so we'll look at how that is used in the context of an ended licence later.

We also spotted we were handling the error incorrectly (passing the error as part of the data arg rather than as its own arg) so fix that as well.

We've also been able to revamp the tests. FetchCurrentReturnCycle, CreateCurrentReturnCycle and GenerateReturnLog all now have their own unit tests so there is no need to retest them here. That includes populating the DB with 'test' record so those services have something to pull. A key benefit of breaking responsibilities down into services is that we can then stub them, and more easily test the behaviour of services that use them.

So, the updated tests are able to replace the need to work with the DB to instead use stubbing.

> TBH, I'm still pondering the stubbing of the insert at this point. It is a simple enough call that it falls under the remit of being 'stubable'. But it also means with my changes we don't have anything which is actually confirming we can persist the data we are generating. I will swing back to this at some point.