Handle individual error in job
Another further tweak we make to the job is handling when a single return log errors. If we are processing 5K return logs, and only intending to run this job per cycle once per year, we don't want a single return log failing to be created blocking the process for the other 4999K return logs.

So, we move the creation part into its own function (and maybe this is where the need for a separate `CreateReturnLogService` becomes a consideration) with error handling.