Demonstrate simplifying FetchReturnCycles
I invested some time in trying to understand what the `FetchReturnCyclesService` was trying to achieve. The conclusion I came to was it is simply fetching all return cycles that end after the date specified.

First, I compared the results of it to a simple query which just applied a `.where('endDate', '>=', date)` clause on a query. Testing each month, starting with today's date and going back more than a year I got matching results.

So, I then swapped the body out and tried the simplified version against the unit tests. They only failed because the order the test was expecting was different. Applying a fixed order and amending the tests to match got them passing as well. No changes were made to what records were expected.

This is obviously not the finished commit, but is just add as a demonstration.