Update ProcessLicenceReturnLogs to create in place
We make a similar change to the service that we did to `ProcessReturnLogs` on the jobs side, and instead of saving the results to an array we then pass to the CreateReturnLogsService, we instead create in place.

You'll notice there is no error handling in this version as well. But more on that in the next change!