Remove all logging and error handling
TBH, I'm still not clear when and how this service will get called. But the two scenarios I think we are looking at is

- when a change happens with a licence
- when [New 'import' job to trigger return log and supplementary flag checks](https://github.com/DEFRA/water-abstraction-system/pull/1453) calls the service

If it errors in the first scenario, we want it to blow up and the user to see it so they are not led to believe that the return logs have been generated. And if we can process 5K return logs in 3 secs, we should be able to handle a single licence in 1 sec, so I feel there is no need for the time to be logged.

In the second instance, we'll be processing all 74K licences. We are only interested in those that have changed (end date added, removed or changed) but we normally see quite a number.

In that case like the return logs job, we're more interested in the time for processing _all_ changed licences, not individual ones. And unlike the return logs job, it is acceptable that a return log creation failing blows up _this_ process, because we are just handling a single licence. It will be on the import job to ensure a licence blowing up doesn't prevent other licences from be processed.