Switch to using licence ID instead of reference
If this is to be triggered from a licence change, in both cases we'll have the licence ID. That would be preferable to use instead of the licence reference so I've updated the `ProcessLicenceReturnLogs` to expect it instead.

This meant updating the `/check` controller, so I also took advantage to make some changes there. I've updated the handler function name to match the route, to be consistent with how we do it elsewhere (even if it is only temporary!)

We also now expect the licence ID as a param rather than in the body, which will make it easier to send the request when testing.