Demonstrate VoidReturnLogsService simplification
Taking insight from the simplification and eventual removal of the `FetchReturnCyclesService` I then looked at `VoidReturnLogsService`.

It was using a very similar logic to the fetch. First, identify the cycle dates for each return cycle (summer and all-year). Then, use this information to query the records.

The conclusion I came to for this was that any return logs that ended before the 'date' of the change, would not be affected by it. If we were to reissue them the engine would spit out the same result.

So, the return logs we need to void and reissue are those that could be affected, and they can be identified by the fact their end date is equal to or after the date change.

With that simplification, the query to both select and patch the relevant returns logs could be done in one.

With this change for example, all the existing unit tests for `VoidReturnLogsService` pass with no changes needed.