Remove VoidReturnLogsService & simplify process
After demonstrating that the void process can be done with one simple query, for now, we don't need the `VoidReturnLogsService` so it has been removed and replaced with a function.

> I'm aware that I am passing through a licence ID and then referring to it as a licence reference. I'm hoping/intending to resolve this in coming changes.

I also think we can simplify what the process service is doing. My impression is that we'd only be getting an 'endDate' for changed licences, but it being optional was to support new licences. With that in mind, there was an expectation new licences would have no return logs so no need to void them.

First, I've changed `endDate` to `changeDate` but it is not a given that we will be receiving just an `endDate`. It could an end date itself has changed, or been removed entirely. So, for now, I;ve renamed it as `changeDate`.

Having done that, it is perfectly valid to treat a new licence's 'change date' as the current date! If we are dealing with a new licence and there are no return logs then calling `voidReturnLogs()` and it doing nothing is not a problem, especially now we have removed all the work that was previously happening.

But I don't trust NALD, so if we are _always_ reissuing return logs for the period, I do not believe it hurts to also ensure we void any existing return logs.

That means we can simplify the process service to always assume it is calling `_voidReturnLogs()`.