Rename fetch requirements
I think this helps make it clear it has a different purpose to the existing fetch service in jobs and helps identify it as being part of the overall process.

I'm going to be tweaking it in the next commit and I'm aware there will still be a lot of duplication. But at this time I'm fine with that.