Refactor FetchLicenceReturnRequirementsService
Sheesh, this went through some changes! This stuff breaks your brain!

Previously, both jobs used the same fetch return requirements service. The optional licence filter we added for testing, came in handy when we needed to start looking at individual licences.

I started out these changes splitting the fetch because I suspected the needs would actually be different. I also envisaged changing the licence fetch to start at the licence level, and then pull the return requirements through as a child record. When I realised I'd end up having to build another generate return logs service, I figured I could stick with the same convention.

The key change is the date used. As seen with my previous changes I believe we can just use the 'change date', rather than having to determine the relevant return cycle dates.

The other major change is _dropping_ the `whereNotIn()` clause that excludes return requirements that already have a return log.

That makes sense (though I'm still mulling it!) when we run the annual jobs because the assumption is we are dealing with licences that have not been processed due to a changes triggering something.

But take a licence that triggers `ProcessLicenceReturnLogsService` because it has been revoked. There will be no new return version, nor return requirements. If we were to run the `FetchLicenceReturnRequirementsService` with that clause in, we would not pick up the existing return requirements.

But we need to, because we need to reissue their return logs with a new end date.

So, I've made the changes I think are necessary, and there is enough complexity that this should be kept as separate beasts.

Beyond that, I've made similar changes to the tests to try and reduce the amount of records generated and try and better elucidate the scenarios the fetch is covering.