Move void logic to back to its own service
So it begins!

What looked to be a simple process before became far more complex as I threw more and more scenarios at the 'engine'. Fetching the return requirements and generating the new return logs is not so bad. But trying to void only those affected by the change took some figuring out.

The previous version identified the matching return cycle for the 'end date' provided, and would then void all return logs linked to it. The engine would then kick in and create the new return logs for the given return cycle.

In some cases, this would lead to it re-generating existing return logs leading to an insert error, which I'm guessing is _actually_ the basis for the 'upsert' query I replaced in the old `CreateReturnLogsService`.

Ideally, we should only be voiding those return logs affected by the change, not all return logs in the _return cycle_ affected by the change.

The solution I came up with was to start with creating the re-issued return logs. When we do, we keep a note of the ID's. We then void the affected return logs afterwards, the criteria being that they

- are for the licence that changed
- are for the return cycle being processed
- have an end date on or after the change date
- do not have a return ID that matches one in `reissuedReturnLogIds`

In my manual testing this did the trick. We could avoid voiding return logs unaffected by the change.

The query to patch the return logs is not that complex, but explaining it and providing unit tests that demonstrate it is. This was enough to warrant putting the feature in its own service.